refactor: move the env id back into its own ids.py module
Keeps the validated-str EnvId, just relocated out of types.py into verifiers/v1/ids.py.
