+++ verifiers/v1/__init__.py
@@ -18,6 +18,7 @@ from verifiers.v1.clients import (
 from verifiers.v1.clients import RolloutContext
 from verifiers.v1.decorators import group_reward, metric, reward, stop
 from verifiers.v1.env import EnvConfig, Environment, TimeoutConfig
+from verifiers.v1.ids import EnvId, ensure_installed, env_name
 from verifiers.v1.retries import RetryConfig
 from verifiers.v1.episode import Episode
 from verifiers.v1.rollout import Rollout
@@ -54,7 +55,6 @@ from verifiers.v1.trace import (
 )
 from verifiers.v1.types import (
     AssistantMessage,
-    EnvId,
     Message,
     Messages,
     Response,
@@ -66,8 +66,6 @@ from verifiers.v1.types import (
     ToolMessage,
     Usage,
     UserMessage,
-    ensure_installed,
-    env_name,
 )
 
 __all__ = [
